pyproject.toml
[project]
name = "stock-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "fastapi>=0.116.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
